Refactored pzsvc library to pull out some broad-use functions into the new "core.go".  Should make it more straightforward to figure out where to go to find things.
